Create gui and db element:
