+++ src/db_customersdata.h
@@ -1,48 +1,76 @@
-#ifndef FM_TRANSACTIONS_WIDGET_H
-#define FM_TRANSACTIONS_WIDGET_H
+#ifndef DB_CUSTOMERS_DATA_H
+#define DB_CUSTOMERS_DATA_H
 
 ////////////////////////////////////////////////////////////////////////////////
 //
 // Includes
 //
-#include "ui_gorcarqner.h"
+#ifndef DB_COMPONENT_H
+#   include "db_component.h"
+#endif
 
-// Qt includes
-#include <QWidget>
+// Qt Includes
+#include <QList>
 ////////////////////////////////////////////////////////////////////////////////
 
 ////////////////////////////////////////////////////////////////////////////////
 namespace fm {
 ////////////////////////////////////////////////////////////////////////////////
-namespace gui {
+namespace db {
 ////////////////////////////////////////////////////////////////////////////////
 
 
 ////////////////////////////////////////////////////////////////////////////////
 //
-// class CTransactionsWidget
+// class CCustomersData
 //
-class CTransactionsWidget : public QWidget
+class CCustomersData : public IDBComponent
 {
 public:// Constructors
-	CTransactionsWidget(QWidget* pwParent = nullptr);
-	~CTransactionsWidget() = default;
+	inline CCustomersData();
+	~CCustomersData() = default;
 
 public:// Interface Methodes
+	void Initialize();
+
+	int GetColumnCount() const;
+	int GetRowCount() const;
+
+	QList<QString> GetCoulmnsName() const;
+
+	QString GetNameByIndex(int nColumn, int nRow) const;
+	void SetNameByIndex(int nColumn, int nRow, QString const& strName);
 
 protected:// Helper Methodes
 
-private:
-	Ui::widget m_uiGorcarqner;
 
+private:// Members
+	int m_nColumnCount;
+	int m_nRowCount;
 };
 ////////////////////////////////////////////////////////////////////////////////
 
 
+
 ////////////////////////////////////////////////////////////////////////////////
-} // namespace gui
+//
+// class CCustomersData
+//
+
+// Constructors
+inline CCustomersData::CCustomersData()
+	: m_nColumnCount(0),
+	  m_nRowCount(0)
+{
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+
+////////////////////////////////////////////////////////////////////////////////
+} // namespace db
 ////////////////////////////////////////////////////////////////////////////////
 } // namespace fm
 ////////////////////////////////////////////////////////////////////////////////
 
-#endif // FM_TRANSACTIONS_WIDGET_H
+#endif // DB_CUSTOMERS_DATA_H
