+++ src/db_storagesdata.cpp
@@ -2,7 +2,7 @@
 //
 // Includes
 //
-#include "fm_storagesdata.h"
+#include "db_storagesdata.h"
 
 // Qt includes
