+++ src/fm_core.h
@@ -37,8 +37,9 @@
 #include <QtCore/qglobal.h>
 
 //  STD includes
-#include <stdexcept>
 #include <new>
+#include <memory>
+#include <stdexcept>
 ////////////////////////////////////////////////////////////////////////////////
 
 ////////////////////////////////////////////////////////////////////////////////
