+++ src/gui_addcustomerdlg.h
@@ -1,5 +1,5 @@
-#ifndef ADD_CUSTOMER_DLG_H
-#define ADD_CUSTOMER_DLG_H
+#ifndef GUI_ADD_CUSTOMER_DLG_H
+#define GUI_ADD_CUSTOMER_DLG_H
 
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -56,4 +56,4 @@ private:
 ////////////////////////////////////////////////////////////////////////////////
 } // namespace fm
 ////////////////////////////////////////////////////////////////////////////////
-#endif // ADD_CUSTOMER_DLG_H
+#endif // GUI_ADD_CUSTOMER_DLG_H
