+++ src/gui_transactionswidget.cpp
@@ -2,7 +2,7 @@
 //
 //	Includes
 //
-#include "fm_transactionswidget.h"
+#include "gui_transactionswidget.h"
 
 // Qt Includes
