+++ src/gui_transactionswidget.h
@@ -1,49 +1,48 @@
-#ifndef FM_TRANSACTIONS_DATA_H
-#define FM_TRANSACTIONS_DATA_H
+#ifndef GUI_TRANSACTIONS_WIDGET_H
+#define GUI_TRANSACTIONS_WIDGET_H
 
 ////////////////////////////////////////////////////////////////////////////////
 //
 // Includes
 //
-#ifndef FM_DB_COMPONENT_H
-#   include "fm_dbcomponent.h"
-#endif
+#include "ui_gorcarqner.h"
 
 // Qt includes
-
+#include <QWidget>
 ////////////////////////////////////////////////////////////////////////////////
 
 ////////////////////////////////////////////////////////////////////////////////
 namespace fm {
 ////////////////////////////////////////////////////////////////////////////////
-namespace db {
+namespace gui {
 ////////////////////////////////////////////////////////////////////////////////
 
 
 ////////////////////////////////////////////////////////////////////////////////
 //
-// class CTransactionsData
+// class CTransactionsWidget
 //
-class CTransactionsData : public IDBComponent
+class CTransactionsWidget : public QWidget
 {
 public:// Constructors
-	inline CTransactionsData() = default;
-	~CTransactionsData() = default;
+	CTransactionsWidget(QWidget* pwParent = nullptr);
+	~CTransactionsWidget() = default;
 
 public:// Interface Methodes
-	void Initialize();
 
 protected:// Helper Methodes
 
+private:
+	Ui::widget m_uiGorcarqner;
 
 };
 ////////////////////////////////////////////////////////////////////////////////
 
 
 ////////////////////////////////////////////////////////////////////////////////
-} // namespace db
+} // namespace gui
 ////////////////////////////////////////////////////////////////////////////////
 } // namespace fm
 ////////////////////////////////////////////////////////////////////////////////
 
-#endif // FM_TRANSACTIONS_DATA_H
+#endif // GUI_TRANSACTIONS_WIDGET_H
